seatserve-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2
